+++ server/test/klite/BrowserTest.kt
@@ -13,5 +13,8 @@ class BrowserTest {
     expect(detectBrowser("Mozilla/5.0+(compatible; UptimeRobot/2.0; http://www.uptimerobot.com/)")).toEqual("UptimeRobot/2.0")
     expect(detectBrowser("Mozilla/5.0 (compatible; Googlebot/2.1; +http://www.google.com/bot.html)")).toEqual("Googlebot/2.1")
     expect(detectBrowser("Mozilla/5.0 (iPhone; CPU iPhone OS 13_2_3 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/13.0.3 Mobile/15E148 Safari/604.1")).toEqual("Mobile/Safari/13.0.3")
+    expect(detectBrowser("Mozilla/5.0 (iPhone; CPU iPhone OS 10_3 like Mac OS X) AppleWebKit/602.1.50 (KHTML, like Gecko) CriOS/56.0.2924.75 Mobile/14E5239e Safari/602.1")).toEqual("Mobile/CriOS/56.0.2924.75")
+    expect(detectBrowser("Mozilla/5.0 (iPhone; CPU iPhone OS 14_4 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/14.0 EdgiOS/46.2.0 Mobile/15E148 Safari/605.1.15")).toEqual("Mobile/EdgiOS/46.2.0")
+    expect(detectBrowser("Mozilla/5.0 (iPhone; CPU iPhone OS 14_4 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) FxiOS/123.0 Mobile/15E148 Safari/605.1.15")).toEqual("Mobile/FxiOS/123.0")
   }
 }